requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
joblib==0.14.1
numpy==1.18.2
pandas==1.0.3
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
